+++ ChangeLog.md
@@ -23,6 +23,7 @@
 - Added a new Event room and a new Merchant room
 - Amethyst Spyglass now reduces scouting costs by 15g per spyglass (and can't reduce past 0g)
 - Scouting Artifact Guardians will now always start with the next upcoming Artifact Guardian
+- Element Swap event now pays 200g instead of costing 100g
 
 #### Prophets of the Labyrinth Version 0.10.0:
 - Added **Consumables**: these resources can be used by any party member during combat at priority speed